+++ src/makeRESTRequest.ts
@@ -1,118 +1,154 @@
-// Derived from https://github.com/intersystems/language-server/blob/bdeea88d1900a3aff35d5ac373436899f3904a7e/server/src/server.ts
+// Derived from
+//  https://github.com/intersystems/language-server/blob/bdeea88d1900a3aff35d5ac373436899f3904a7e/server/src/server.ts
 
-import axios, { AxiosResponse } from 'axios';
-import axiosCookieJarSupport from 'axios-cookiejar-support';
-import tough = require('tough-cookie');
-import { ServerSpec } from './extension';
+import axios, { AxiosResponse } from "axios";
+import axiosCookieJarSupport from "axios-cookiejar-support";
+import tough = require("tough-cookie");
+import * as vscode from "vscode";
+import { AUTHENTICATION_PROVIDER } from "./authenticationProvider";
+import { IServerSpec } from "./extension";
 
 axiosCookieJarSupport(axios);
 
 /**
  * Cookie jar for REST requests to InterSystems servers.
  */
-let cookieJar: tough.CookieJar = new tough.CookieJar();
+export const cookieJar: tough.CookieJar = new tough.CookieJar();
 
-export interface AtelierRESTEndpoint {
-    apiVersion: number,
-    namespace: string,
-    path: string
-};
+export interface IAtelierRESTEndpoint {
+    apiVersion: number;
+    namespace: string;
+    path: string;
+}
 
 /**
  * Make a REST request to an InterSystems server.
- * 
+ *
  * @param method The REST method.
  * @param server The server to send the request to.
  * @param endpoint Optional endpoint object. If omitted the request will be to /api/atelier/
  * @param data Optional request data. Usually passed for POST requests.
  */
- export async function makeRESTRequest(method: "HEAD"|"GET"|"POST", server: ServerSpec, endpoint?: AtelierRESTEndpoint, data?: any): Promise<AxiosResponse | undefined> {
+export async function makeRESTRequest(
+    method: "HEAD"|"GET"|"POST",
+    server: IServerSpec,
+    endpoint?: IAtelierRESTEndpoint,
+    data?: any,
+    ): Promise<AxiosResponse | undefined> {
 
-	// Build the URL
-	var url = server.webServer.scheme + "://" + server.webServer.host + ":" + String(server.webServer.port);
+    // Build the URL
+    let url = server.webServer.scheme + "://" + server.webServer.host + ":" + String(server.webServer.port);
     const pathPrefix = server.webServer.pathPrefix;
-	if (pathPrefix && pathPrefix !== "") {
-		url += pathPrefix;
-	}
+    if (pathPrefix && pathPrefix !== "") {
+        url += pathPrefix;
+    }
     url += "/api/atelier/";
     if (endpoint) {
         url += "v" + String(endpoint.apiVersion) + "/" + endpoint.namespace + endpoint.path;
     }
 
-	// Make the request (SASchema support removed)
-	try {
-        var respdata: AxiosResponse;
+    // Make the request (SASchema support removed)
+    try {
+        let respdata: AxiosResponse;
         if (data !== undefined) {
             // There is a data payload
             respdata = await axios.request(
                 {
-                    method: method,
-                    url: encodeURI(url),
-                    data: data,
+                    data,
                     headers: {
-                        'Content-Type': 'application/json'
+                        "Content-Type": "application/json",
+                    },
+                    jar: cookieJar,
+                    method,
+                    url: encodeURI(url),
+                    validateStatus: (status) => {
+                        return status < 500;
                     },
                     withCredentials: true,
-                    jar: cookieJar,
-                    validateStatus: function (status) {
-                        return status < 500;
-                    }
-                }
+                },
             );
-            if (respdata.status === 401 && typeof server.username !== 'undefined' && typeof server.password !== 'undefined') {
-                // Either we had no cookies or they expired, so resend the request with basic auth
-
-                respdata = await axios.request(
-                    {
-                        method: method,
-                        url: encodeURI(url),
-                        data: data,
-                        headers: {
-                            'Content-Type': 'application/json'
+            if (respdata.status === 401) {
+                // Use AuthenticationProvider to get password if not supplied by caller
+                await resolveCredentials(server);
+                if (typeof server.username !== "undefined" && typeof server.password !== "undefined") {
+                    // Either we had no cookies or they expired, so resend the request with basic auth
+                    respdata = await axios.request(
+                        {
+                            auth: {
+                                password: server.password,
+                                username: server.username,
+                            },
+                            data,
+                            headers: {
+                                "Content-Type": "application/json",
+                            },
+                            jar: cookieJar,
+                            method,
+                            url: encodeURI(url),
+                            withCredentials: true,
                         },
-                        auth: {
-                            username: server.username,
-                            password: server.password
-                        },
-                        withCredentials: true,
-                        jar: cookieJar
-                    }
-                );
+                    );
+                }
             }
-        }
-        else {
+        } else {
             // No data payload
             respdata = await axios.request(
                 {
-                    method: method,
-                    url: encodeURI(url),
-                    withCredentials: true,
                     jar: cookieJar,
-                    validateStatus: function (status) {
+                    method,
+                    url: encodeURI(url),
+                    validateStatus: (status) => {
                         return status < 500;
-                    }
-                }
+                    },
+                    withCredentials: true,
+                },
             );
-            if (respdata.status === 401 && typeof server.username !== 'undefined' && typeof server.password !== 'undefined') {
-                // Either we had no cookies or they expired, so resend the request with basic auth
-
-                respdata = await axios.request(
-                    {
-                        method: method,
-                        url: encodeURI(url),
-                        auth: {
-                            username: server.username,
-                            password: server.password
+            if (respdata.status === 401) {
+                // Use AuthenticationProvider to get password if not supplied by caller
+                await resolveCredentials(server);
+                if (typeof server.username !== "undefined" && typeof server.password !== "undefined") {
+                    // Either we had no cookies or they expired, so resend the request with basic auth
+                    respdata = await axios.request(
+                        {
+                            auth: {
+                                password: server.password,
+                                username: server.username,
+                            },
+                            jar: cookieJar,
+                            method,
+                            url: encodeURI(url),
+                            withCredentials: true,
                         },
-                        withCredentials: true,
-                        jar: cookieJar
-                    }
-                );
+                    );
+                }
             }
         }
         return respdata;
-	} catch (error) {
-		console.log(error);
-		return undefined;
-	}
-};
+    } catch (error) {
+        console.log(error);
+        return undefined;
+    }
+}
+
+export async function resolveCredentials(serverSpec: IServerSpec) {
+    // This arises if setting says to use authentication provider
+    if (typeof serverSpec.password === "undefined") {
+        const scopes = [serverSpec.name, serverSpec.username || ""];
+        let session = await vscode.authentication.getSession(
+            AUTHENTICATION_PROVIDER,
+            scopes,
+            { silent: true },
+        );
+        if (!session) {
+            session = await vscode.authentication.getSession(
+                AUTHENTICATION_PROVIDER,
+                scopes,
+                { createIfNone: true },
+            );
+        }
+        if (session) {
+            serverSpec.username = session.scopes[1] === "UnknownUser" ? "" : session.scopes[1];
+            serverSpec.password = session.accessToken;
+        }
+    }
+}